requirements.txt
aiogram>=3.0
aiohttp
pyyaml
pydantic
pytz
supabase
dotenv